Added Maintenance Mode to database
This includes a function in datastore for a key/value store of
persistent state.  So far, it's only being used to persist the
maintenance mode across restarts.  It is not resolved how this
state will be flagged when the database is split up into monthly
files (each time the state is changed, this is noted in the log,
but one may need go go back through several files to find out what
state the detector is in at any point in time).